+++ DeepTimeLib/IAgent.cs
@@ -2,11 +2,11 @@
 
 public interface IAgent<TState, TAction>
 {
-    void SetNext(TState state, float reward);
+    void SetNext(TState state, double reward);
     TAction Eval();
 
     void StartEpisode(TState state);
-    void EndEpisode(TState state, float reward);
+    void EndEpisode(TState state, double reward);
 
     string ToJson();
 }
